+++ src/main/java/dev/wefhy/whymap/config/RenderConfig.kt
@@ -11,6 +11,7 @@ object RenderConfig {
 //    }
 
     private val ignoredOverlayBlocks = arrayOf(
+        "air",
         "string",
         "tripwire",
         "vine",