+++ src/main/java/dev/wefhy/whymap/config/WhyUserSettings.kt
@@ -3,8 +3,6 @@
 package dev.wefhy.whymap.config
 
 import dev.wefhy.whymap.WhyMapClient
-import dev.wefhy.whymap.gui.WhyConfirmScreen
-import dev.wefhy.whymap.libs.whysettings.CustomSetSettingsEntry
 import dev.wefhy.whymap.libs.whysettings.SettingsEntry
 import dev.wefhy.whymap.libs.whysettings.SettingsEntry.Companion.addColorPicker
 import dev.wefhy.whymap.libs.whysettings.SettingsEntry.Companion.addSlider
@@ -13,9 +11,6 @@ import dev.wefhy.whymap.libs.whysettings.WhySettings
 import dev.wefhy.whymap.libs.whysettings.WhySettingsCategory
 import dev.wefhy.whymap.whygraphics.WhyColor
 import dev.wefhy.whymap.whygraphics.intARGB
-import kotlinx.coroutines.GlobalScope
-import kotlinx.coroutines.launch
-import net.minecraft.client.MinecraftClient
 
 object WhyUserSettings: WhySettings() {
     val generalSettings = GeneralSettingsCategory().register()
@@ -28,7 +23,6 @@ object WhyUserSettings: WhySettings() {
         mapSettings.minimapPosition = userSettings.minimapPosition
         mapSettings.minimapMode = userSettings.minimapMode
         serverSettings.exposeHttpApi = userSettings.exposeHttpApi
-        mapSettings.forceExperimentalMinmap = userSettings.forceExperimentalMinmap
         generalSettings.hudColor = WhyColor.fromARGB(userSettings.hudColor)
     }
 
@@ -39,7 +33,6 @@ object WhyUserSettings: WhySettings() {
             minimapPosition = mapSettings.minimapPosition,
             minimapMode = mapSettings.minimapMode,
             exposeHttpApi = serverSettings.exposeHttpApi,
-            forceExperimentalMinmap = mapSettings.forceExperimentalMinmap,
             hudColor = generalSettings.hudColor.intARGB
         )
     }
@@ -51,26 +44,26 @@ class GeneralSettingsCategory : WhySettingsCategory("General") {
 }
 
 class MapSettingsCategory: WhySettingsCategory("Map") {
-    var forceExperimentalMinmap by CustomSetSettingsEntry(
-        false
-    ) { proposedValue, actuallySet ->
-        if (proposedValue) {
-            GlobalScope.launch {
-                with(MinecraftClient.getInstance()) {
-                    WhyConfirmScreen(
-                        "Experimental minimap",
-                        "This minimap is experimental and WILL cause crashes if used more than a few minutes. Are you sure you want to enable it?"
-                    ) {
-                        if (it) {
-                            actuallySet()
-                        }
-                    }.show()
-                }
-            }
-        } else {
-            actuallySet()
-        }
-    }.addToggle("Force Minimap (experimental, will cause crashes)")
+//    var forceExperimentalMinmap by CustomSetSettingsEntry(
+//        false
+//    ) { proposedValue, actuallySet ->
+//        if (proposedValue) {
+//            GlobalScope.launch {
+//                with(MinecraftClient.getInstance()) {
+//                    WhyConfirmScreen(
+//                        "Experimental minimap",
+//                        "This minimap is experimental and WILL cause crashes if used more than a few minutes. Are you sure you want to enable it?"
+//                    ) {
+//                        if (it) {
+//                            actuallySet()
+//                        }
+//                    }.show()
+//                }
+//            }
+//        } else {
+//            actuallySet()
+//        }
+//    }.addToggle("Force Minimap (experimental, will cause crashes)")
     var minimapPosition by SettingsEntry(UserSettings.MinimapPosition.TOP_LEFT).addToggle("Minimap position")
     var minimapMode by SettingsEntry(WhyMapClient.MapMode.NORTH_LOCKED).addToggle("Minimap mode")
     var mapScale by SettingsEntry(1.0).addSlider("Map scale", 0.5, 2.0)